requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0,<0.31.0
SQLAlchemy>=2.0.0,<2.1.0
pydantic>=2.6.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
mysqlclient>=2.2.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
alembic>=1.13.0,<2.0.0
reflex>=0.4.0,<0.6.0
python-jose[cryptography]>=3.3.0,<4.0.0
bcrypt>=4.0.0,<5.0.0
